requirements.txt
Django>=4.2
djangorestframework
django-environ
django-cors-headers
psycopg2-binary
gunicorn
django-allauth
djangorestframework-simplejwt
